modified:   .config/sway/config 	modified:   .config/sway/config.d/behavior.conf 	modified:   .config/sway/config.d/keybindings.conf 	new file:   .config/swayidle/config 	new file:   .config/waybar/colors/colors.css 	modified:   .config/waybar/config 	modified:   .config/waybar/style.css 	modified:   .config/zsh/.zprofile 	modified:   .config/zsh/.zshenv 	modified:   .config/zsh/.zshrc 	new file:   .local/bin/logout.sh 	modified:   .local/bin/statusbar/audio_input.sh 	modified:   .local/bin/statusbar/audio_output.sh 	modified:   .local/bin/statusbar/backlight.sh 	modified:   .local/bin/statusbar/battery.py 	modified:   .local/bin/statusbar/date.sh 	new file:   .local/bin/statusbar/statusbar.sh 	deleted:    .local/bin/statusbar/waybar.sh 	deleted:    Pictures/Wallpapers/desktop-bg.jpg 	modified:   packages.txt

+++ .local/bin/statusbar/audio_input.sh
@@ -1,28 +1,46 @@
 #!/bin/sh
+# Audio input (microphone) module for statusbar
+# NOTE: Only one process (per session) of this script is allowed!
 
 value=5
 scontrol='Capture'
 script_name="${0##*/}"
-pipe="$XDG_RUNTIME_DIR/bar-${script_name%.*}-pipe"
+runtime_dir="$XDG_RUNTIME_DIR/statusbar-modules-$XDG_SESSION_ID"
+pipe="$runtime_dir/${script_name%.*}-pipe"
+
+if [ -p "$pipe" ]; then
+	printf 'This program is already running!\n' 1>&2
+	exit 1
+fi
+
+[ -d "$runtime_dir" ] || mkdir -m 700 "$runtime_dir"
 mkfifo -m 600 "$pipe"
 
-while true; do
-	audio_info="$(amixer sget "$scontrol" | tail -1)"
-	volume="$(printf '%s' "$audio_info" | grep -oE '[0-9]+%' | tr -d '%')"
+_exit() { printf 'exit\n' > "$pipe"; }
+trap _exit INT TERM
 
-	if printf '%s' "$audio_info" | grep -q '\[off\]'; then
-		output="<span fgcolor=\"#E79D78\"> $volume%</span>"
-	elif [ "$volume" -gt 100 ]; then
-		output="<span fgcolor=\"#DC657D\"> $volume%</span>"
-	else output=" $volume%"; fi;
+{
+	while true; do
+		audio_info="$(amixer sget "$scontrol" | tail -1)"
+		volume="$(printf '%s' "$audio_info" | grep -oE '[0-9]+%' | tr -d '%')"
+		class=''
 
-	printf '%s\n' "$output"
+		if printf '%s' "$audio_info" | grep -q '\[off\]'; then
+			text=" $volume%"
+			class='muted'
+		elif [ "$volume" -gt 100 ]; then
+			text=" $volume%"
+			class='exceeded'
+		else text=" $volume%"; fi
 
-	read -r mes < "$pipe"
-	case $mes in
-		mute) amixer -q sset "$scontrol" toggle;;
-		inc) amixer -q sset "$scontrol" ${value}%+;;
-		dec) amixer -q sset "$scontrol" ${value}%-;;
-        exit) rm -f "$pipe"; exit;;
-	esac
-done
+		printf '{"text": "%s", "class": "%s"}\n' "$text" "$class"
+
+		read -r mes < "$pipe"
+		case $mes in
+			mute) amixer -q sset "$scontrol" toggle;;
+			inc) amixer -q sset "$scontrol" ${value}%+;;
+			dec) amixer -q sset "$scontrol" ${value}%-;;
+			exit) rm -f "$pipe"; exit;;
+		esac
+	done
+} & wait
